Move path to tree to core and make internal
